[core] Introduce the TaskScheduler
This is the class that will own all the schedulers we use
during the lifetime of a Map. The Map has to hold a shared_ptr
in order to use the static members of the class, otherwise
it will return invalid references.

This class will make sure to cleanup all the schedulers when
they are no longer needed.